Rename ui/msg/{msg -> view}
