+++ src/ui/msg/view/interact.ts
